Update sessions_manager.php
When starting a new session, the sessions manager now checks for and deletes all other sessions associated with the same UUID on the same host. UUID is used as a condition so multiple users can authenticate on a single IP.

Cookies fixed so all pages on the site can see them.

Started isValidSession() framework